moving dispatchers into a file in reporters rather than it's own place to fit with theme and support dynamic module loading
